Restructure changes to reduce the diff
